// src/app/base/main/dashboard/grid/@enums/widget.enum.ts
export enum WIDGETS {
  chart = 'chart',
  columnChart = 'columnChart',
  rapidPay = 'rapidPay',
  table = 'table',
  cards = 'cards',
  calender = 'calender',
  advertisement = 'advertisement',
  advertisement1 = 'advertisement1',
  currentAccountDetails = 'currentAccountDetails',
  creditCardDetails = 'creditCardDetails',
  depositDetails = 'depositDetails',
  loanDetails = 'loanDetails',
  quickLinks = 'quickLinks',
  videos = 'videos',
  accountStatement = 'accountStatement',
  pendingAuthorizationTable = 'pendingAuthorizationTable',
}